backend/legacy_reference/app/requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
flatbuffers==22.12.6
fonttools==4.38.0
frozendict==2.3.4
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multitasking==0.0.11
numpy==1.24.0
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.3
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
yfinance==0.2.3
motor==3.1.1
